Interrupts example code
Here is some example code for an interrupt driven program. I use two
buttons which move a servo just a little bit in either direction. I also
use the tinyprintf module and Uart in order to print to the console what
button is pressed. I was going to use this as a model for the gps and
other sensors.